fix: removing get_doc dependencies for lms course
