+++ school/lms/doctype/lms_course_review/lms_course_review.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021, FOSS United and contributors
+# Copyright (c) 2021, Frappe and contributors
 # For license information, please see license.txt
 
 import frappe
