+++ school/public/css/style.css
@@ -10,6 +10,7 @@
   --text-3xl: 22px;
   --text-4xl: 44px;
   --navbar-shadow: 0px 1px 8px rgba(0, 0, 0, 0.08)
+  --gray-750: #505A62;
 }
 
 input[type=checkbox] {
@@ -56,15 +57,16 @@ input[type=checkbox] {
 .course-card-pills {
   background: #ffffff;
   margin-left: 0;
-  margin-right: .5rem;
-  border-radius: 4px;
-  padding: 4px 6px;
-  font-size: 10px;
+  margin-right: 1rem;
+  border-radius: 6px;
+  padding: 3.5px 8px;
+  font-size: 11px;
   text-align: center;
   letter-spacing: 0.011em;
   text-transform: uppercase;
-  font-weight: bold;
-  box-shadow: var(--shadow-base);
+  font-weight: 600;
+  box-shadow: var(--popover-box-shadow);
+  color: var(--gray-900);
 }
 
 .dark-pills {
@@ -94,6 +96,8 @@ input[type=checkbox] {
 
 .course-card {
   flex-direction: column;
+  height: 100%;
+  min-height: 350px;
 }
 
 .muted-text {
@@ -102,17 +106,24 @@ input[type=checkbox] {
 }
 
 .course-card-meta {
-  margin: 16px 0px 8px;
+  margin: 0.75rem 0 0.5rem;
+  font-size: 14px;
+}
+
+.course-card-meta-2 {
+  color: var(--gray-750);
 }
 
 .course-card-content {
-  width: 100%;
-  padding: 0px 1rem 1.25rem;
+  padding: 0 1.25rem 1.25rem;
+  display: flex;
+  flex-direction: column;
+  flex: 1 1 auto;
 }
 
 @media (max-width: 350px) {
   .course-card-content {
-    padding: 0px 10px 20px;
+    padding: 0px 10px 10px;
   }
 }
 
@@ -124,8 +135,8 @@ input[type=checkbox] {
 
 .course-card-title {
   font-size: 1.125rem;
-  margin-bottom: 1.5rem;
-  height: 56px;
+  font-weight: 600;
+  margin-bottom: 1.25rem;
 }
 
 @media (max-width: 360px) {
@@ -145,16 +156,19 @@ input[type=checkbox] {
 }
 
 .course-instructor {
-  margin-left: 8px;
-  font-size: 12px;
-  line-height: 135%;
-  color: var(--text-color-dark);
+  margin-left: 0.625rem;
+  font-size: 0.875rem;
 }
 
 .course-student-count {
   display: flex;
   font-size: 12px;
   float: right;
+  font-weight: 500;
+}
+
+.course-card-footer {
+  margin-top: auto;
 }
 
 .view-course-link {
@@ -1181,7 +1195,7 @@ input[type=checkbox] {
 
 .progress {
   width: 100%;
-  height: 8px;
+  height: 4px;
 }
 
 .progress-bar {
@@ -1189,12 +1203,9 @@ input[type=checkbox] {
 }
 
 .progress-percentage {
-  width: 100%;
-  font-size: 12px;
-  line-height: 165%;
-  letter-spacing: 0.02em;
-  color: #000000;
-  text-align: center;
+  margin: 0.5rem 0 1.3rem;
+  font-size: 0.8rem;
+  font-weight: 500;
 }
 
 pre {
