+++ school/www/courses/index.py
@@ -6,21 +6,21 @@ def get_context(context):
     context.live_courses, context.upcoming_courses = get_courses()
     context.restriction = check_profile_restriction()
     context.metatags = {
-        "title": "All Courses",
+        "title": "All Live Courses",
         "image": frappe.db.get_single_value("Website Settings", "banner_image"),
         "description": "This page lists all the courses published on our website",
         "keywords": "All Courses, Courses, Learn"
     }
 
 def get_courses():
-    course_names = frappe.get_all("LMS Course",
+    courses = frappe.get_all("LMS Course",
                                 filters={"is_published": True},
-                                fields=["name", "upcoming"])
+                                fields=["name", "upcoming", "title", "image"])
 
     live_courses, upcoming_courses = [], []
-    for course in course_names:
+    for course in courses:
         if course.upcoming:
-            upcoming_courses.append(frappe.get_doc("LMS Course", course.name))
+            upcoming_courses.append(course)
         else:
-            live_courses.append(frappe.get_doc("LMS Course", course.name))
+            live_courses.append(course)
     return live_courses, upcoming_courses
